refactor(config-ui): rename toast2 to toast
