+++ config-ui/src/components/index.ts
@@ -22,6 +22,6 @@ export * from './page-header'
 export * from './selector'
 export * from './dialog'
 export * from './table'
-export * from './toast2'
+export * from './toast'
 export * from './logo'
 export * from './card'
